+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.1] - 2022-02-16
+
 ### Changed
 
 -   When game updates happen, Hephaistos will now try to patch anyway (with warnings) instead of not patching (with errors).
@@ -312,7 +314,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Initial release.
 
-[Unreleased]: https://github.com/nbusseneau/hephaistos/compare/v1.6.0...HEAD
+[Unreleased]: https://github.com/nbusseneau/hephaistos/compare/v1.6.1...HEAD
+
+[1.6.1]: https://github.com/nbusseneau/hephaistos/compare/v1.6.0...v1.6.1
 
 [1.6.0]: https://github.com/nbusseneau/hephaistos/compare/v1.5.1...v1.6.0